requirements/base.txt
fastapi
pydantic-settings
uvicorn
tortoise-orm[asyncpg]
aerich
